% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_user.R
\name{create_dir_if_doesnt_exist}
\alias{create_dir_if_doesnt_exist}
\title{Check that directory exists}
\usage{
create_dir_if_doesnt_exist(dir_libros, year)
}
\description{
Check that directory exists
}
