% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_libro_gastos.R
\name{create_libro_gastos}
\alias{create_libro_gastos}
\title{create libro de gastos}
\usage{
create_libro_gastos(
  db_gastos,
  dir_libros,
  users = c("Andres", "Martina", "Pulpo"),
  year = 2021,
  ...
)
}
\arguments{
\item{db_gastos}{A tibble with the gastos. This data base is returned by \code{pulpo_clean_gastos()}}

\item{dir_libros}{A directory path to export the books}

\item{users}{A string. Name for which the book will be created. Options: c("Andres", "Martina", "Pulpo")}
}
\description{
create libro de gastos
}
