% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulpo_clean_gastos.R
\name{pulpo_clean_gastos}
\alias{pulpo_clean_gastos}
\title{Clean gastos report and prepare variables for Libro de gastos}
\usage{
pulpo_clean_gastos(db, db_receptores, year, ...)
}
\arguments{
\item{db}{A tibble. Raw data of gastos report as it is exported from Zoho}

\item{db_receptores}{A tibble. Raw data of receptores report as it is exported from Zoho}

\item{year}{A number. year for which the report is to be created}
}
\value{
A tibble, with all the variables for libro de gastos
}
\description{
Clean gastos report and prepare variables for Libro de gastos
}
