% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulpo_clean_invoices.R
\name{pulpo_clean_invoices}
\alias{pulpo_clean_invoices}
\title{Clean invoices report}
\usage{
pulpo_clean_invoices(db = invoices, year)
}
\arguments{
\item{db}{A tibble. Data base with the raw version of the invoice report. To get this
report, run \code{pulpo_get_reportes()}}

\item{year}{A number. Year for which the libro de gastos is to be generated}
}
\value{
A tibble with all the relevant variables to create libro de gastos
}
\description{
Cleans invoice report to create relevant variables for \code{libro_de_ingresos}
}
