% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulpo_get_reportes.R
\name{pulpo_get_reportes}
\alias{pulpo_get_reportes}
\title{Download reports to from zoho creator}
\usage{
pulpo_get_reportes(zoho_password, zoho_token)
}
\arguments{
\item{zoho_password}{A string. Password of araupontones in zoho creator}

\item{zoho_token}{A string. Token of Zoho creator API of araupontone}
}
\value{
A list with the reports of: \code{invoices}, \code{gastos}, \code{receptores},
\code{expenses}
}
\description{
Download reports to from zoho creator
}
